refactor Python class bench, proxy, peer copy
